Adds a comparison function and passing comparison test
In order to sort cards, there needs to be a way to evaluate their
relative values given suit ranking (for games where there are suit
orders, like hearts or bridge). This method takes two cards and
assigns a modifier based on the value of the suit (passed in), and
adds the modifier to the face value of the cards. High cards are
converted to numerical values using a dict look-up.

Tests comparison for two game suit ranking systems: alphabetical
(used in bridge), and alternating (used in some Chinese games).
The value of the clubs and diamonds should alternate based on the
ranking, both should be beaten by spades regardless. Passing.